M6h: fix executor fallback to preserve run counters; accept [[items]] + f64 weights
While validating M6h against the real Anspire proxy
(qwen36 1 RPS / 60s / 2-concurrency multi-turn), the saved
`metrics.json` showed `scheduled: 0` / `session_count: 0`
even when the run had actually scheduled 6 ticks and
completed 6 sessions (18 turns). The HTML report's
"Prompt cache" section worked, but the counters that the
executor maintained independently of the per-request
records were silently dropped on the
`Arc::try_unwrap` fallback path.

M6h also accepts the user-facing TOML aliases
`[[items]]` (instead of the M6-introduced `[[prompt]]`)
and `weight = 1.0` floats (instead of `weight = 1`
integers). Both changes are non-breaking — the original
shapes still work.

What changed
------------
* `src/runner/metrics.rs`:
  - New `MetricsAggregator::merge_counters_from(&mut self,
    other: &Self)`. The executor's fallback path now calls
    it after rebuilding the aggregator from
    `per_request`. It carries over `scheduled`,
    `skipped`, `session_count`, and `session_dropped` —
    the four counters that `push_record_clone` doesn't
    already cover (`session_turn_total` and the
    prompt-cache totals ARE covered by `push_record_clone`
    via `accumulate_cache` and the per-record branch,
    so the merge deliberately leaves them alone to avoid
    double-counting).
  - New test `merge_counters_from_carries_over_bookkeeping`
    exercises the full record → push_record_clone →
    merge sequence and verifies the no-double-count
    invariant.
* `src/runner/executor.rs`: the `Arc::try_unwrap`
  fallback branch now calls `fresh.merge_counters_from(
  &guard)` after replaying per-request records.
* `src/dataset/custom.rs`:
  - `ProfileFile.prompt` gains `#[serde(alias = "items")]`
    so user-facing `[[items]]` table-arrays are accepted
    alongside the M6-introduced `[[prompt]]`. The example
    templates in `docs/examples/datasets/` use
    `[[items]]` because "items" reads more naturally for
    multi-message conversation profiles (where the entries
    aren't really "prompts" in the literal sense).
  - `ProfileItem.weight` is now `Option<f64>` instead of
    `Option<u32>`. Users can write `weight = 1.0` /
    `weight = 0.5` / `weight = 2.5` without thinking
    about integer vs float. The internal `DatasetItem.weight`
    is also `f64`. No code actually does weighted
    sampling today, so the change is forward-looking.
* `src/dataset/mod.rs`: `DatasetItem.weight` annotation
  updated to reflect the `f64` switch.
* `tests`: `custom.rs` existing test
  `parses_toml_single_turn` updated to assert on `Some(3.0)`
  instead of `Some(3)`. New test `toml_items_alias_accepted`
  covers the `[[items]]` schema alias.

Real-Anspire validation
-----------------------
Smoke test (qwen36 single-turn 1 RPS / 30s):
- 31/31 success, 0 errors
- p50 latency 244ms, TTFT p50 203ms
- history at `<root>/qwen36/<run-id>/`
- summary shows `alias: qwen36` line

Multi-turn test (qwen36 dynamic_multi 1 RPS / 60s / 2
parallel sessions, 3 turns × 3 conversations via the
`docs/examples/datasets/multi-turn-conversation.toml`
template):
- 18/18 success, 0 errors (6 sessions × 3 turns)
- achieved 0.30 RPS (target 1.00; qwen36 ~3 RPS rate limit
  on the ollama-anspire provider is the bottleneck)
- `metrics.json`: `session_count: 6`,
  `session_turn_total: 18`, `session_dropped: 0`,
  `scheduled: 6`, `skipped: 55` — all correct
- `prompt cache hit rate: 16.96%` (turn 1 20.27%, turn 2+
  16.81%) — qwen36 reports a small amount of
  `cached_tokens` (token-level overlap between sessions
  using the same conversation item); the cache_creation
  fields are 0 because the ollama-anspire provider doesn't
  surface Anthropic-style cache_creation
- HTML report: Model card shows `qwen36`, Configuration
  table has a "Model alias: qwen36 (groups history +
  compare)" row, Prompt cache section with three bar
  rows (Overall / Turn 1 / Turn 2+) — all working

Tests
-----
* 189 unit + 12 e2e = 201 passing (was 199, +2 new unit
  tests: `merge_counters_from_carries_over_bookkeeping`
  and `toml_items_alias_accepted`).
* 0 warnings, `#![deny(warnings)]` holds.
* No new external deps.

Backward compat
---------------
* `DatasetItem.weight` and `ProfileItem.weight` change
  type from `u32` to `f64`. Integer values (`weight = 3`)
  still parse correctly (3 → 3.0). Old `config.json`
  files that never used `weight` are unaffected.
* `ProfileFile.prompt` gets a `serde(alias = "items")`
  for the deserialization path; the on-disk format
  hasn't changed. The previous `[[prompt]]` form
  continues to work.
* `metrics.json` files from M1-M6g load via
  `from_records`, which calls `push_record_clone` (the
  only path that existed before M6h). They don't go
  through `merge_counters_from` because they're
  round-tripping per-request records, not the live
  executor state. The only consumer of the
  fallback path is the executor itself.

+++ src/dataset/mod.rs
@@ -32,9 +32,11 @@ pub struct DatasetItem {
     pub prompt: String,
     /// Rough pre-flight estimate of prompt token count.
     pub estimated_prompt_tokens: u32,
-    /// Optional sampling weight. `None` or `0` items are skipped. `1`
-    /// is the default when not set.
-    pub weight: Option<u32>,
+    /// Optional sampling weight. `None` or `0` items are
+    /// skipped. `1.0` is the default when not set. `f64` so
+    /// users can write `0.5` / `1.0` / `2.5` in TOML profiles
+    /// without thinking about integer vs float.
+    pub weight: Option<f64>,
     /// Optional tags. Carried through for report grouping.
     pub tags: Vec<String>,
     /// Optional stable name. Defaults to a synthetic id.
